# Use struct for non-nullable tickets