--- a/pkg/document/internal_document.go
+++ b/pkg/document/internal_document.go
@@ -120,14 +120,14 @@ func (d *InternalDocument) ApplyChangePack(pack *change.Pack) error {
 	d.checkpoint = d.checkpoint.Forward(pack.Checkpoint)
 
 	if pack.MinSyncedTicket != nil {
-		d.GarbageCollect(pack.MinSyncedTicket)
+		d.GarbageCollect(*pack.MinSyncedTicket)
 	}
 
 	return nil
 }
 
 // GarbageCollect purge elements that were removed before the given time.
-func (d *InternalDocument) GarbageCollect(ticket *time.Ticket) int {
+func (d *InternalDocument) GarbageCollect(ticket time.Ticket) int {
 	return d.root.GarbageCollect(ticket)
 }