<!--  Thanks for sending a pull request! -->

**What this PR does / why we need it**:

Use a struct instead of pointer: Ticket

Modified to use struct for non-nullable tickets.

**Which issue(s) this PR fixes**:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Fixes #


It has been confirmed that performance is slightly lower when using struct in text.

Before: https://github.com/yorkie-team/yorkie/runs/5086458392?check_suite_focus=true
```
go test -tags bench -benchmem -bench=. ./test/bench
goos: linux
goarch: amd64
pkg: github.com/yorkie-team/yorkie/test/bench
cpu: Intel(R) Xeon(R) Platinum 8370C CPU @ 2.80GHz
BenchmarkDocument/constructor_test-2         	  681069	      1875 ns/op	     848 B/op	      15 allocs/op
BenchmarkDocument/status_test-2              	 1191621	       918.9 ns/op	     816 B/op	      13 allocs/op
BenchmarkDocument/equals_test-2              	  119319	     10497 ns/op	    5760 B/op	     114 allocs/op
BenchmarkDocument/nested_update_test-2       	   38211	     31059 ns/op	   13210 B/op	     338 allocs/op
BenchmarkDocument/delete_test-2              	   30632	     38714 ns/op	   15866 B/op	     399 allocs/op
BenchmarkDocument/object_test-2              	   85519	     14025 ns/op	    6592 B/op	     134 allocs/op
BenchmarkDocument/array_test-2               	   23938	     51220 ns/op	   14978 B/op	     400 allocs/op
BenchmarkDocument/text_test-2                	   28760	     40568 ns/op	   13434 B/op	     422 allocs/op
BenchmarkDocument/text_composition_test-2    	   28330	     42000 ns/op	   18042 B/op	     489 allocs/op
BenchmarkDocument/rich_text_test-2           	    8846	    139505 ns/op	   46001 B/op	    1351 allocs/op
BenchmarkDocument/counter_test-2             	   27414	     43597 ns/op	   14474 B/op	     415 allocs/op
BenchmarkDocument/text_edit_gc_100-2         	     124	   9513628 ns/op	 3178076 B/op	   81130 allocs/op
BenchmarkDocument/text_edit_gc_1000-2        	       2	 883677688 ns/op	298640928 B/op	 7514829 allocs/op
BenchmarkDocument/text_split_gc_100-2        	     100	  10439209 ns/op	 3711330 B/op	   81575 allocs/op
BenchmarkDocument/text_split_gc_1000-2       	       2	 936290228 ns/op	364663672 B/op	 7523082 allocs/op
BenchmarkDocument/text_100-2                 	    3793	    323769 ns/op	  107498 B/op	    4788 allocs/op
BenchmarkDocument/text_1000-2                	     345	   3538356 ns/op	 1027175 B/op	   47091 allocs/op
BenchmarkDocument/array_1000-2               	     363	   3216571 ns/op	 1610898 B/op	   36398 allocs/op
BenchmarkDocument/array_10000-2              	      32	  35283586 ns/op	15360816 B/op	  370251 allocs/op
BenchmarkDocument/array_gc_100-2             	    3336	    366763 ns/op	  165564 B/op	    3905 allocs/op
BenchmarkDocument/array_gc_1000-2            	     295	   4008854 ns/op	 1827322 B/op	   45265 allocs/op
BenchmarkDocument/counter_1000-2             	    2647	    455855 ns/op	  239832 B/op	    8540 allocs/op
BenchmarkDocument/counter_10000-2            	     248	   4861585 ns/op	 2905054 B/op	   89549 allocs/op
BenchmarkDocument/rich_text_100-2            	    3444	    354013 ns/op	  131288 B/op	    4547 allocs/op
BenchmarkDocument/rich_text_1000-2           	     320	   3795378 ns/op	 1250765 B/op	   44150 allocs/op
BenchmarkDocument/object_1000-2              	     285	   4131397 ns/op	 2039077 B/op	   36429 allocs/op
BenchmarkDocument/object_10000-2             	      21	  49542247 ns/op	19072846 B/op	  370746 allocs/op
2022-02-06T23:56:31.029Z	INFO	default	etcd connected, URI: [localhost:2379]
BenchmarkSync/memory_sync_10_test-2          	  166665	      7136 ns/op	    1101 B/op	      29 allocs/op
BenchmarkSync/memory_sync_100_test-2         	   20480	     58033 ns/op	    6635 B/op	     195 allocs/op
BenchmarkSync/memory_sync_1000_test-2        	    1971	    631441 ns/op	   60301 B/op	    1727 allocs/op
BenchmarkSync/memory_sync_10000_test-2       	     190	   6755378 ns/op	  638043 B/op	   18084 allocs/op
BenchmarkSync/etcd_sync_100_test-2           	1000000000	         0.3006 ns/op	       0 B/op	       0 allocs/op
BenchmarkTextEditing-2                       	       1	64405006102 ns/op	20547264024 B/op	447734689 allocs/op
```

After: https://github.com/yorkie-team/yorkie/runs/5092193950?check_suite_focus=true

```
go test -tags bench -benchmem -bench=. ./test/bench
goos: linux
goarch: amd64
pkg: github.com/yorkie-team/yorkie/test/bench
cpu: Intel(R) Xeon(R) Platinum 8272CL CPU @ 2.60GHz
BenchmarkDocument/constructor_test-2         	  628672	      1593 ns/op	     864 B/op	      15 allocs/op
BenchmarkDocument/status_test-2              	 1328878	      1075 ns/op	     832 B/op	      13 allocs/op
BenchmarkDocument/equals_test-2              	  111189	      9258 ns/op	    5880 B/op	     113 allocs/op
BenchmarkDocument/nested_update_test-2       	   41913	     28683 ns/op	   13658 B/op	     332 allocs/op
BenchmarkDocument/delete_test-2              	   33279	     35993 ns/op	   16394 B/op	     394 allocs/op
BenchmarkDocument/object_test-2              	   94874	     12757 ns/op	    6784 B/op	     134 allocs/op
BenchmarkDocument/array_test-2               	   25861	     46830 ns/op	   15554 B/op	     395 allocs/op
BenchmarkDocument/text_test-2                	   32386	     36870 ns/op	   14082 B/op	     412 allocs/op
BenchmarkDocument/text_composition_test-2    	   30500	     39592 ns/op	   19410 B/op	     471 allocs/op
BenchmarkDocument/rich_text_test-2           	    8750	    128792 ns/op	   47505 B/op	    1333 allocs/op
BenchmarkDocument/counter_test-2             	   32754	     36620 ns/op	   15178 B/op	     391 allocs/op
BenchmarkDocument/text_edit_gc_100-2         	     123	   9612078 ns/op	 3459262 B/op	   90533 allocs/op
BenchmarkDocument/text_edit_gc_1000-2        	       2	 901776328 ns/op	323056388 B/op	 8508877 allocs/op
BenchmarkDocument/text_split_gc_100-2        	     100	  10260963 ns/op	 3981973 B/op	   91373 allocs/op
BenchmarkDocument/text_split_gc_1000-2       	       2	 956459070 ns/op	388973632 B/op	 8521116 allocs/op
BenchmarkDocument/text_100-2                 	    3594	    327135 ns/op	  124530 B/op	    4390 allocs/op
BenchmarkDocument/text_1000-2                	     327	   3779483 ns/op	 1195407 B/op	   43093 allocs/op
BenchmarkDocument/array_1000-2               	     396	   2993891 ns/op	 1682963 B/op	   35397 allocs/op
BenchmarkDocument/array_10000-2              	      30	  34886607 ns/op	16081217 B/op	  360251 allocs/op
BenchmarkDocument/array_gc_100-2             	    3516	    341460 ns/op	  172966 B/op	    3805 allocs/op
BenchmarkDocument/array_gc_1000-2            	     325	   3731587 ns/op	 1899637 B/op	   44265 allocs/op
BenchmarkDocument/counter_1000-2             	    2822	    425490 ns/op	  263920 B/op	    7539 allocs/op
BenchmarkDocument/counter_10000-2            	     268	   4424467 ns/op	 3145141 B/op	   79548 allocs/op
BenchmarkDocument/rich_text_100-2            	    3435	    353501 ns/op	  148456 B/op	    4149 allocs/op
BenchmarkDocument/rich_text_1000-2           	     316	   3779964 ns/op	 1419134 B/op	   40152 allocs/op
BenchmarkDocument/object_1000-2              	     302	   3920395 ns/op	 2143240 B/op	   37427 allocs/op
BenchmarkDocument/object_10000-2             	      22	  50327887 ns/op	20115416 B/op	  380751 allocs/op
2022-02-07T11:24:28.942Z	INFO	default	etcd connected, URI: [localhost:2379]
BenchmarkSync/memory_sync_10_test-2          	  172804	      6531 ns/op	    1101 B/op	      29 allocs/op
BenchmarkSync/memory_sync_100_test-2         	   20998	     58038 ns/op	    6632 B/op	     195 allocs/op
BenchmarkSync/memory_sync_1000_test-2        	    1641	    694316 ns/op	   58652 B/op	    1627 allocs/op
BenchmarkSync/memory_sync_10000_test-2       	     166	   7177524 ns/op	  606979 B/op	   16828 allocs/op
BenchmarkSync/etcd_sync_100_test-2           	1000000000	         0.3034 ns/op	       0 B/op	       0 allocs/op
BenchmarkTextEditing-2                       	       1	75113485165 ns/op	22061126944 B/op	508174603 allocs/op
```

**Special notes for your reviewer**:

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".
-->
```release-note

```

**Additional documentation**:

<!--
This section can be blank if this pull request does not require a release note.

Please use the following format for linking documentation:
- [Usage]: <link>
- [Other doc]: <link>
-->
```docs

```

**Checklist**:
- [ ] Added relevant tests or not required
- [ ] Didn't break anything